+++ border-py-gym-env/src/tch.rs
@@ -1,11 +1,10 @@
 mod ndarray_act;
 mod ndarray_converter;
 mod ndarray_obs;
-mod tensor_batch;
 pub use ndarray_act::NdarrayAct;
 pub use ndarray_converter::{NdarrayConverter, NdarrayConverterConfig};
 pub use ndarray_obs::NdarrayObs;
-pub use tensor_batch::{TensorBatch, ZeroTensor};
+pub use border_generic_replay_buffer::tch::TensorBatch;
 
 use ndarray::ArrayD;
 use num_traits::cast::AsPrimitive;