+++ test_before_pr.sh
@@ -1,10 +1,17 @@
+#!/usr/bin/env bash
+# Stop immediately if any test command fails.
+set -e
+
 # crates
 cargo test -p border-core
+cargo test -p border-generic-replay-buffer --features candle,tch
 cargo test -p border-py-gym-env
+cargo test -p border-py-gym-env --features candle
 cargo test -p border-py-gym-env --features tch
 cargo test -p border-async-trainer
 cargo test -p border-atari-env
 cargo test -p border-candle-agent
+cargo test -p border-minari --features candle
 cargo test -p border-tch-agent
 cargo test -p border-policy-no-backend --features=tch